@@ -1,7 +1,6 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery
-
-  helper :all # include all helpers, all the time+
+  
   helper_method :fix_host
 
 
@@ -12,5 +11,5 @@ class ApplicationController < ActionController::Base
       end
     return content
    end
-
+   
 end
