@@ -1,17 +1,6 @@
-# Use Bundler (preferred)
-begin
-  require File.expand_path('../../.bundle/environment', __FILE__)
-rescue LoadError
-  require 'rubygems'
-  require 'bundler'
-  Bundler.setup
+require 'rubygems'
 
-  # To use 2.x style vendor/rails and RubyGems
-  #
-  # vendor_rails = File.expand_path('../../vendor/rails', __FILE__)
-  # if File.exist?(vendor_rails)
-  #   Dir["#{vendor_rails}/*/lib"].each { |path| $:.unshift(path) }
-  # end
-  #
-  # require 'rubygems'
-end
+# Set up gems listed in the Gemfile.
+ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
+
+require 'bundler/setup' if File.exists?(ENV['BUNDLE_GEMFILE'])
