@@ -1,23 +1,24 @@
-# This file is auto-generated from the current state of the database. Instead of editing this file, 
-# please use the migrations feature of Active Record to incrementally modify your database, and
-# then regenerate this schema definition.
+# This file is auto-generated from the current state of the database. Instead
+# of editing this file, please use the migrations feature of Active Record to
+# incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your database schema. If you need
-# to create the application database on another system, you should be using db:schema:load, not running
-# all the migrations from scratch. The latter is a flawed and unsustainable approach (the more migrations
+# Note that this schema.rb definition is the authoritative source for your
+# database schema. If you need to create the application database on another
+# system, you should be using db:schema:load, not running all the migrations
+# from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20100217153556) do
+ActiveRecord::Schema.define(:version => 20110522064103) do
 
   create_table "feed_urls", :force => true do |t|
     t.string   "feed_url"
     t.string   "title"
+    t.string   "site_url"
     t.boolean  "star",       :default => false
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "site_url"
   end
 
   create_table "feeds", :force => true do |t|
